feat(asr): delete the retired XPC ASR helper and every consumer that existed only for it (#1908 chunk C)
Chunk A+B (already merged onto this branch) made ASRManager the sole
ASRManagerInterface implementation; ASRManagerProxy and
Sources/EnviousWisprASRService/ were never constructed after that, but
still existed on disk. This deletes them and everything that existed
solely to serve the XPC path they made possible, per the approved plan's
Chunk C.

Deleted outright:
- Sources/EnviousWisprASR/ASRManagerProxy.swift (the XPC client)
- Sources/EnviousWisprASRService/ (the XPC helper target — 5 files)
- Sources/EnviousWisprCore/ASRServiceProtocol.swift (the shared wire protocol)
- Sources/EnviousWisprCore/XPCOperationSignalFile.swift (the XPC-only
  signal/watchdog file mechanism — but see below, one type from this file
  is relocated, not deleted)
- Sources/EnviousWisprAppKit/App/DictationRuntime/ASREventRouter.swift (its
  entire body was wiring asrManager.onServiceInterrupted, which is gone)
- ASRManagerInterface.onServiceInterrupted (protocol requirement + both
  conformers' storage)
- KernelDictationDriver.handleASRServiceInterruption() and
  RecordingSessionKernel.externalASRInterrupted() — the App-routed XPC-crash
  entry points. RecordingSessionKernel.routeASRInterruption(sid:), the
  internal mechanism they called, stays — it has a separate, still-live
  production caller (adapter.onEngineInterrupted, an in-process
  engine-crash signal unrelated to XPC).
- The entire "#959 warm-respawn" subsystem: KernelDictationDriver
  .residentModelLostWhileIdle / .warmRespawnInFlight / .warmRespawnStartedAt
  / .beginWarmRespawnOverlay() / .markResidentModelLostWhileIdle() /
  .updateWarmRespawnLatch(), ColdPressGuard.resolveNotReadyPress()'s
  warm-respawn branch (the function's ONLY branch that ever returned
  `true` — traced through both RecordingStarter.swift call sites to
  confirm neither can reach a warm-respawn outcome anymore), and
  TelemetryService.serviceReclaimed/.serviceRespawnStarted/
  .serviceRespawnCompleted. This entire mechanism existed only to give a
  press a fast re-warm path specifically after the OS reaped the idle XPC
  helper — the scenario #1908's whole point is to make impossible. (The
  *general* warm-respawn UI concept — showing the recording pill instead
  of a caching flash during a brief re-warm — has no other trigger in this
  codebase; it was never a general mechanism, despite living on general
  names.)
- XPCASRTransportError entirely (all 7 cases; every throw site lived in
  the two now-deleted files) — including its two remaining production
  consumers: ParakeetEngineAdapter's one-shot stale-helper transport retry
  (loadModelWithTransportRecovery(), now just a direct asrManager.loadModel()
  call) and retryDecode()'s stale-readiness-mirror repair branch (an
  XPC-round-trip-specific staleness the in-process manager cannot have,
  since isModelLoaded never goes stale independently of a real connection
  object in-process).
- The four ASRManagerProxy*Tests suites, ASREventRouterTests,
  ASREventRouterCeilingsTests, RecordingSessionKernelExternalInterruptionTests,
  and the KernelDictationDriverBridgeMatrixTests
  handleASRServiceInterruption matrix section — each tested exactly the
  retired mechanism above and nothing else.
- ASRManagerProxy/EnviousWisprASRService's Package.swift and Project.swift
  target definitions, dependencies, and settings.

Relocated (kept, not deleted):
- OneShotContinuationASR: lived in ASRManagerProxy.swift but is used by
  ASRManager.performLoadRacingCancellation(); moved into ASRManager.swift.
- XPCOperationSignalWedgeError: lived in XPCOperationSignalFile.swift but
  is a Sentry-identity pin with real historical production issues riding
  on its exact string (ENVIOUSWISPR-22, ENVIOUSWISPR-1B) — it already had
  no live producer before this change (its one prior producer,
  AudioCaptureProxy.swift, was deleted before #1908 even started). Moved
  to its own file, Sources/EnviousWisprCore/XPCOperationSignalWedgeError.swift,
  for the same reason it survived that earlier deletion: preserving a
  shipped identity string, never change once shipped.

Adapted (kept the property, changed the trigger):
- Four kernel salvage/retry test files (RecordingSessionKernelSalvageTests,
  KernelPhase2RetryTests, KernelSalvageRetryTests, and the shared
  Simulator/ScenarioRunner.swift) drove the salvage/retry machinery under
  test via kernel.externalASRInterrupted() as one of several ways to reach
  it. That machinery (routeASRInterruption(sid:) and everything downstream
  of it — Phase-2 retry, the salvage ladder, .asrInterrupted as a
  RecordingOutcome) is NOT XPC-specific and stays fully live; only the
  trigger changed, to context.engine.fireEngineInterrupted() (an existing
  FakeEngine method already built for exactly this substitution, wired
  through the adapter's own still-live onEngineInterrupted callback).

DebugFaultEndpoint's asrProxy dependency and its force_xpc_kill command
are removed (nothing left to kill); AppLifecycleCoordinator's construction
site updated to match.

Not touched in this commit: scripts/build-dev-app.sh, build-release-dmg.sh,
.github/workflows/release.yml, scripts/ci/assert-app-launches.sh,
scripts/check-xpc-error-hygiene.sh + its pr-check.yml step, Tests/RuntimeUAT/,
and the docs update (architecture.md/capability-map.md/pipeline-mechanics.md).
Those are the plan's remaining Chunk C/D scope, landing as follow-up
commits on this same branch before the PR opens (per the plan's §12
rollback correction: this ships as one complete refactor, not
separately-mergeable pieces).

Full test suite green: 7,559 cases (down from 7,638 — net effect of
deleting several whole suites built around the retired mechanism, no
proportional new-test count expected for a deletion chunk). Two
frozen-inventory tests (EngineMutationInventoryFreezeTests,
EngineProtocolSurfaceFreezeTests) updated for the removed call sites and
protocol member, each reclassified individually against live source, not
batch-approved. scripts/test-inventory-baseline.txt regenerated (7 stale
grandfather entries for the deleted suites removed).

Self-audit-grep: grep -rln "ASRManagerProxy\|ASRServiceHandler\|EnviousWisprASRService\|useXPCASRService\|XPCASRTransportError\|ASRServiceProtocol" Sources/ Tests/EnviousWisprASRTests Tests/EnviousWisprTests Package.swift Project.swift (every remaining hit checked individually — comments describing accurate history, or the two deliberately-relocated types; no dangling code reference)
Self-audit-owner: docs/feature-requests/issue-1908-2026-09-07-xpc-collapse-parakeet-in-process.md
