+++ Tests/EnviousWisprTests/Architecture/HeartPathBreadcrumbLiteralsTests.swift
@@ -8,14 +8,16 @@ import Testing
 /// or removes one.
 ///
 /// AppLogger heart-path log prefixes are per-router (`[AudioEventRouter]` /
-/// `[ASREventRouter]` / `[WedgeRecoveryRouter]`) per founder decision Q4
-/// 2026-05-19. The per-router prefixes are asserted present here. A regression
-/// that re-merges the prefixes is caught by the epic #763 freeze test, which
+/// `[WedgeRecoveryRouter]`) per founder decision Q4 2026-05-19. The
+/// per-router prefixes are asserted present here. A regression that
+/// re-merges the prefixes is caught by the epic #763 freeze test, which
 /// scans all of `Sources/` for the deleted root-state type's name.
+/// (#1908: `ASREventRouter.swift` itself, and its `[ASREventRouter]` prefix,
+/// were deleted along with the XPC path it bridged — see the two NOTEs below
+/// for the same pattern on earlier deletions.)
 @Suite struct HeartPathBreadcrumbLiteralsTests {
   private static let routerSourcePaths = [
     "Sources/EnviousWisprAppKit/App/DictationRuntime/AudioEventRouter.swift",
-    "Sources/EnviousWisprAppKit/App/DictationRuntime/ASREventRouter.swift",
     "Sources/EnviousWisprAppKit/App/DictationRuntime/WedgeRecoveryRouter.swift",
   ]
 
@@ -35,12 +37,15 @@ import Testing
   // audio-capture boundary itself — those breadcrumbs can no longer be produced
   // in-process. The engine-interruption breadcrumb survives, now spelled "Audio
   // engine interrupted."
+  // NOTE (#1908): the `[ASREventRouter] ASR onServiceInterrupted` literal was
+  // removed with `ASREventRouter.swift` itself — the last XPC helper
+  // collapsed in-process, so `asrManager.onServiceInterrupted` (the only
+  // thing that router ever wired) has no live producer anywhere.
   private static let requiredLiterals: [String] = [
     "\"Audio engine interrupted\"",
     "\"parakeet_state\"",
     "\"whisperkit_state\"",
     "[AudioEventRouter] Audio onEngineInterrupted",
-    "[ASREventRouter] ASR onServiceInterrupted",
   ]
 
   @Test func requiredLiteralsPresent() throws {
