+++ scripts/test-inventory-baseline.txt
@@ -7,10 +7,6 @@
 Tests/EnviousWisprASRTests/ASRManagerBackendInjectionTests.swift	ASRManagerBackendInjectionTests
 Tests/EnviousWisprASRTests/ASRManagerBackendInjectionTests.swift	ASRManagerLoadGenerationTests
 Tests/EnviousWisprASRTests/ASRManagerColdStateContractTests.swift	ASRManagerColdStateContractTests
-Tests/EnviousWisprASRTests/ASRManagerProxyCancellationTests.swift	ASRManagerProxyCancellationTests
-Tests/EnviousWisprASRTests/ASRManagerProxyLoadCompletionTests.swift	ASRManagerProxyLoadCompletionTests
-Tests/EnviousWisprASRTests/ASRManagerProxyProgressPollingTests.swift	ASRManagerProxyProgressPollingTests
-Tests/EnviousWisprASRTests/ASRManagerProxyTranscribeCompletionTests.swift	ASRManagerProxyTranscribeCompletionTests
 Tests/EnviousWisprASRTests/EngineMutationScopeTests.swift	EngineMutationScopeTests
 Tests/EnviousWisprASRTests/FluidAudioBridgeClassificationTests.swift	FluidAudioBridgeClassificationTests
 Tests/EnviousWisprASRTests/LanguageDetectorTests.swift	LanguageDetectorTests
@@ -30,7 +26,6 @@ Tests/EnviousWisprTests/ASR/ParakeetStreamingSentryErrorTests.swift	ParakeetStre
 Tests/EnviousWisprTests/ASR/ParakeetTranscriptionSentryErrorTests.swift	ParakeetTranscriptionSentryErrorTests
 Tests/EnviousWisprTests/ASR/WhisperKitModelLoadSentryErrorTests.swift	WhisperKitModelLoadSentryErrorTests
 Tests/EnviousWisprTests/ASR/WhisperKitSetupServiceTests.swift	WhisperKitSetupServiceTests
-Tests/EnviousWisprTests/App/ASREventRouterTests.swift	ASREventRouterTests
 Tests/EnviousWisprTests/App/AccessibilityWarmupObserverTests.swift	AccessibilityWarmupObserverTests
 Tests/EnviousWisprTests/App/AppLoggerLaunchSyncTests.swift	AppLoggerLaunchSyncTests
 Tests/EnviousWisprTests/App/AppWindowCoordinatorTests.swift	AppWindowCoordinatorTests
@@ -99,7 +94,6 @@ Tests/EnviousWisprTests/App/TranscriptCoordinatorTests.swift	TranscriptCoordinat
 Tests/EnviousWisprTests/App/UpdateCoordinatorProactiveCheckTests.swift	UpdateCoordinatorProactiveCheckTests
 Tests/EnviousWisprTests/App/WedgeRecoveryRouterTests.swift	WedgeRecoveryRouterTests
 Tests/EnviousWisprTests/App/WhisperPreviewSessionHandleTests.swift	WhisperPreviewSessionHandleTests
-Tests/EnviousWisprTests/Architecture/ASREventRouterCeilingsTests.swift	ASREventRouterCeilingsTests
 Tests/EnviousWisprTests/Architecture/AdapterShapeTests.swift	AdapterShapeTests
 Tests/EnviousWisprTests/Architecture/AppDelegateCeilingsTests.swift	AppDelegateCeilingsTests
 Tests/EnviousWisprTests/Architecture/AppDelegateOwnershipTests.swift	AppDelegateOwnershipTests
@@ -360,7 +354,6 @@ Tests/EnviousWisprTests/Pipeline/RecordStartTelemetrySinkTests.swift	RecordStart
 Tests/EnviousWisprTests/Pipeline/RecordingCapInvariantTests.swift	RecordingCapInvariantTests
 Tests/EnviousWisprTests/Pipeline/RecordingSessionContinuingIdentityTests.swift	RecordingSessionContinuingIdentityTests
 Tests/EnviousWisprTests/Pipeline/RecordingSessionKernelDeadAirFloorTests.swift	RecordingSessionKernelDeadAirFloorTests
-Tests/EnviousWisprTests/Pipeline/RecordingSessionKernelExternalInterruptionTests.swift	RecordingSessionKernelExternalInterruptionTests
 Tests/EnviousWisprTests/Pipeline/RecordingSessionKernelSalvageTests.swift	EngineInterruptionCauseRecoverabilityTests
 Tests/EnviousWisprTests/Pipeline/RecordingSessionKernelSalvageTests.swift	ExhaustedRetrySpoolDeletionTests
 Tests/EnviousWisprTests/Pipeline/RecordingSessionKernelSalvageTests.swift	InterruptionReasonTests
